Let mobile users click
The click event works for other areas of the program. I have moved the Mouse Dow and Mouse Up detection outside into their own rules and provides a result for a new variable "drag". If the user drags the mouse it will se to True, else will be False.

Only when it is false will the click event continue. This should resolve the issue for mobile users being unable to click on the canvas.