Adds support for mobile correctly
Mobile users cant now select points, add new coordinates and interact with the map.

When users click on a point it will also highlight it in the list and scroll to it making the feature work. The original feature for PC users to hover the mouse over the points for the same thing is still in the code.

I have also touched the z-index for the list and for the "are you sure you want to delete" screen so they do not overlap incorrectly