List is now updated when deleting points
Fixed the issue where the list was not updating when a point is deleted. To resolve the issue I have asses the updatePointList(); to the redrawCanvas(); so that it is udated every time the canvas is. I have also removed the original calls so it does not run twice and simplifies the code

Added Breakaway Large for TDM

Added support for both PNG and JPEG files

  const supportedExtensions = ["png", "jpg", "jpeg"];
  const selectedExtension = supportedExtensions.find((ext) =>
    checkImageExists(`Map Images/${selectedGameMode}/${selectedMapOption}.${ext}`)
  );

  if (selectedExtension) {
    const imagePath = `Map Images/${selectedGameMode}/${selectedMapOption}.${selectedExtension}`;
  } else {
    console.log("Image not found for the selected options.");
  }